+++ src/Razor/src/Microsoft.VisualStudio.LanguageServerClient.Razor/Extensions/PositionExtensions.cs
@@ -5,10 +5,9 @@ using System;
 using Microsoft.AspNetCore.Razor.LanguageServer.RazorLS;
 using Microsoft.CodeAnalysis.Text;
 using Microsoft.Extensions.Logging;
-using OmniSharp.Extensions.LanguageServer.Protocol.Models;
-using VS = Microsoft.VisualStudio.LanguageServer.Protocol;
+using Microsoft.VisualStudio.LanguageServer.Protocol;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Extensions
+namespace Microsoft.VisualStudio.LanguageServerClient.Razor.Extensions
 {
     internal static class PositionExtensions
     {
