+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/JsonRpc/JsonRpcRequestSchedulerTest.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Razor.Test.Common;
 using Microsoft.VisualStudio.Threading;
 using OmniSharp.Extensions.JsonRpc;
 using Xunit;
