+++ src/Razor/test/Microsoft.VisualStudio.LanguageServerClient.Razor.Test/TestLogger.cs
@@ -1,0 +1,30 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT license. See License.txt in the project root for license information.
+
+#nullable disable
+
+using System;
+using Microsoft.Extensions.Logging;
+
+namespace Microsoft.VisualStudio.LanguageServerClient.Razor.Test
+{
+    internal class TestLogger : ILogger
+    {
+        public static readonly TestLogger Instance = new();
+
+        public IDisposable BeginScope<TState>(TState state)
+        {
+            return default;
+        }
+
+        public bool IsEnabled(LogLevel logLevel)
+        {
+            return true;
+        }
+
+        public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception exception, Func<TState, Exception, string> formatter)
+        {
+            // noop
+        }
+    }
+}
